Factor out Sherlock Crosselling
